fix: Post module

+++ modules/Post/Controllers/PostController.php
@@ -2,8 +2,8 @@
 
 namespace Modules\Post\Controllers;
 
-use Illuminate\Routing\Controller;
-use Modules\Post\Models\Post;
+use App\Http\Controllers\Controller;
+use App\Models\Post;
 use Modules\Post\Requests\Store;
 use Modules\Post\Requests\Update;
 use Modules\Post\Tables\PostTableView;
@@ -12,28 +12,38 @@ class PostController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', Post::class); // Otorisasi
+
         return PostTableView::make()->view('post::index');
     }
 
     public function create()
     {
+        $this->authorize('create', Post::class); // Otorisasi
+
         return view('post::create');
     }
 
     public function store(Store $request)
     {
-        Post::create($request->validated());
+        $post = Post::create($request->validated());
 
-        return to_route('modules::post.index')->withSuccess('Post saved');
+        return to_route('modules::post.show', $post)->withSuccess('Post saved');
     }
 
     public function show(Post $post)
     {
+        $this->authorize('show', $post); // Otorisasi
+
+        $post->load(['topic', 'writer']);
+
         return view('post::show', compact('post'));
     }
 
     public function edit(Post $post)
     {
+        $this->authorize('update', $post); // Otorisasi
+
         return view('post::edit', compact('post'));
     }
 
@@ -41,11 +51,13 @@ class PostController extends Controller
     {
         $post->update($request->validated());
 
-        return to_route('modules::post.index')->withSuccess('Post saved');
+        return to_route('modules::post.show', $post)->withSuccess('Post saved');
     }
 
     public function destroy(Post $post)
     {
+        $this->authorize('delete', $post); // Otorisasi
+
         $post->delete();
 
         return to_route('modules::post.index')->withSuccess('Post deleted');
